\name{IMak}
\alias{IMak}
\alias{IMak}
\docType{package}
\title{Item Maker}
\description{This is an Automatic Item Generator for Psychological Assessment. Items created with the IMak package should not be used in applied settings as part of the working protocol without ensuring first that the items meet the required psychometric quality standards.}
\details{
\tabular{ll}{
Package: \tab IMak\cr
Type: \tab Package\cr
Version: \tab 1.2.1\cr
Date: \tab 2018-08-06\cr
License: \tab GPL-3\cr
}
This package includes functions for building information about items and then plotting them. Functions \code{build_fa} and \code{plot_fa} are currently available. \code{build_fa} creates an object of class \code{'fa_items'} which contains information about figural analogies. \code{plot_fa} plots the latter object.
}
\author{
Diego Blum

Maintainer: Diego Blum <blumworx@gmail.com>
}
\references{
Blum, D., & Holling, H. (2018). Automatic generation of figural analogies with the IMak package. \emph{Frontiers in psychology, 9}(1286), 1-13. DOI:10.3389/fpsyg.2018.01286
}

\keyword{IMak}
\seealso{
\code{\link{build_fa}}
\code{\link{plot_fa}}
}
