\name{IMak}
\alias{IMak}
\alias{IMak}
\docType{package}
\title{Item Maker}
\description{This is an Automatic Item Generator for Psychological Testing. It is recommended for research purposes only.}
\details{
\tabular{ll}{
Package: \tab IMak\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2016-01-20\cr
License: \tab GPL-3\cr
}
This package includes functions for building information about items and then plotting them. Functions \code{build_fa} and \code{plot_fa} are currently available. \code{build_fa} creates an object of class \code{'fa_items'} which contains information about figural analogies. \code{plot_fa} plots the latter object.
}
\author{
Diego Blum

Maintainer: Diego Blum <blumworx@gmail.com>
}
\references{
Blum, D., Holling, H., Galibert, M.S., & Forthmann, B. (2016, in review). Task difficulty prediction of figural analogies. \emph{Intelligence}.
}

\keyword{IMak}
\seealso{
\code{\link{build_fa}}
\code{\link{plot_fa}}
}
